Subtle event listener tweak to avoid unnecessary term change notifications.
The input control was triggering events during its initialization that subsequently triggered extra (inaccurate) events notifying `<radium-filter-panel>` listeners that the terms had changed.

This caused our application to think the user had changed the filter critieria when they actually hadn’t, such that we incorrectly marked our saved filters as dirty and triggered unnecessary extra service calls.

In `radium-filter-panel::_createInputControl()`, it is listening for `value-changed`, which fires whenever the value changes. However, we’re only interested in when the user made changes to the value.

I switched this to listen to `change` instead. From the Polymer docs for `<paper-input>`:

> change - Fired when the input changes due to user interaction.

Now this behaves like the listeners used for the other filter control types, and only triggers in response to a user gesture.