Added “clear” event to disambiguate between changes and clearing an autocomplete.
When the `radium-filter-autocomplete` is cleared, it dispatches a `change` event (just like if the field is otherwise edited via typing).

We are using the `radium-filter-autocomplete` component in a standalone manner within a form in our application, and need to know when the user specifically *cleared* the UI (either via the clear button or failing to autocomplete to an available suggestion).

Added logic to fire `clear` events for these operations.