Fix to refit() dropdown if previously fit to loading indicator’s height.
If the autocomplete source takes long enough that the loading indicator is shown, the dropdown will be fit to the height of the loading indicator. When suggestions were subsequently applied, the dropdown was not being refit to new bounds.

This adds a call to refit() the dropdown once suggestions arrive (scheduled to occur after the browser reflow from rendering the suggestions).