Reworked support for hidden controls.
Instead of a using a `type` of `hidden`, filters now have an optional `hidden` property. If set to `true`, the controll will still be rendered to the filter panel, but it will be hidden via CSS via `display: none`. This way, filter controls that dynamically look up labels from values (ex. autocomplete / autocompletelist) will still do so when restoring values from the querystring.